Fundamental re-design of ROMineVolume, including fixes to several bugs, from minor to critical:
https://github.com/DarklightGames/DarkestHour/issues/598 (all enemy players inside a MV when it activates are (1) initially ignore & (2) all die as soon as another enemy enters the MV!)
Solution summary: when MV is activated, check for enemies already inside & if any are present then record a MineAreaEnterTime for them & set a PainTimer.

https://github.com/DarklightGames/DarkestHour/issues/599 (players who enter a mine volume in a vehicle & then exit are instantly killed)
Solution summary: when a vehicle enters the MV or we find a vehicle in the MV, loop through all occupants & record MineAreaEnterTime for all player pawns.

https://github.com/DarklightGames/DarkestHour/issues/600 (players in a mine volume who enter a vehicle or switch vehicle position can prolong their 'life' inside the MV)
Solution summary: check pawn's DrivenVehicle is none before accepting it as valid Touch & recording MineAreaEnterTime (pawn will have a DrivenVehicle if just entered or switched)

https://github.com/DarklightGames/DarkestHour/issues/465 (vehicle passengers do not get minefield warnings for vehicles in a minefield)
Solution summary: when a vehicle enters the MV or we find a vehicle in the MV, loop through all occupants & give a warning to any players.

Also paves the way for a much enhanced spawn protection mine volume in a subclass.