Big change to the `DH_ATWeapons` package:
I've always found the package distinction between AT weapons and non-AT weapons nonsensical. It makes about as much sense as separating bolt-action rifles into a separate package because they "behave differently" than other weapons. As far as I can tell, there's no reason to put them into another package other than as an organizational step that leads to a slippery slope of over-classification.

This change transplants all the classes from `DH_ATWeapons` to `DH_Weapons`. However, to maintain backwards compatibility with existing custom maps or maps that explicitly change the weapon load-out of roles, we need to maintain a bunch of classes that are likely to be used. As a result, classes that are likely to need a persistent reference are now subclasses of the "real" class in the `DH_Weapons` package.

For example, there is a `DH_BazookaWeapon` in `DH_Weapons`, but to maintain backwards compatibility with the `DH_ATWeapons.DH_BazookaWeapon` class, the `DH_ATWeapons.DH_BazookaWeapon` class just subclasses the `DH_Weapons.DH_BazookaWeapon` class.

The `DH_ATWeapons` package will be completely removed in v8.