1st stage of fix for https://github.com/DarklightGames/DarkestHour/issues/406 (artillery damage precedes explosion emitter). Deprecated confusing & badly written ROArtilleryShell class. DHArtilleryShell now contains everything it did and directly extends the base Projectile class. This will also kick start a transition to having a base DHProjectile parent class, with common functionality. Renamed several variables and functions for clarity (one bad example: bAlreadyPlayedFarSound flagged that the CLOSE sound had already been played, not the DISTANT sound!). The basic functionality is the same, but have refactored all of it for clarity, as overall it was pretty obfuscated - now it is easy to follow.
Includes fix for problems/confusion in RO class re making the shell visible, stemming from the fact that an arty shell can spawn several metres above the ground a second or 2 before it actually begins to fall.
That's because if the close/incoming sound is longer than the shell's flight time to impact, the shell is 'held' and not dropped until its impact will coincide with the end of the close sound.
So if the shell is made visible at that point you see a hanging shell.
RO 1st tried hiding the shell with bVisible=false, but that stopped it replicating.
So they ended up hacking it by setting the DrawScale to 0.001 so the hanging shell couldn't be seen.
I have simply started the shell with a DrawType of none, then set the DrawType to StaticMesh when the shell gets dropped from its hanging position (so it is invisible until then).
Having a DrawType of none makes it invisible but doesn't prevent replication of the shell actors to net clients.
And as the incoming shell is now visible I have updated it from a panzerfaust warhead to a large shell.
Now, if you are looking directly at an incoming arty strike, you will just see a the large HE shells flashing down to impact.