Improvement on fix for https://github.com/DarklightGames/DarkestHour/issues/639 (Brecourt can be made unwinnable if the Axis satchel Gun 3 or 4 before they are active), as discussed. In the destroyable mesh class I've replaced bDestroyableByAxis and bDestroyableByAllies with a single TeamCanDamage enum. Defaults to the enum zero value of 'both', but leveller can specify the mesh can only be damaged by allies or axis.
Updated Brecourt map to use new system (same as the demo map except for the 'can damage' variables).
Set all four gun DSMs so they can only be damaged by the allies.
Set guns 3 & 4 to be initially deactivated.
Added two DSM modify actors to activate the DSMs for guns 3 & 4 when guns 1 & 2 are both destroyed.
Renamed the event that is triggered when guns 1 & 2 are both destroyed  from 'Deactivate_Gun3and4Prot' to 'ActivateGuns3And4' to be more generic.
That event now deactivates the mine volume protecting guns 3 & 4 from the allies in the earlier stage, plus now activated the DSMs for guns 3 & 4.
Full rebuild with lighting to RGB8.
Run through the map and all seems to behave as desired.
Deleted the temp demo map I did for Brecourt.