Massive re-factor, without any functional changes:
* "Mortar Observers" are now named "Artillery Observers".
* All references to "mortar targets" have been replaced with "artillery targets", in comments, strings, variable names etc.

I realize this creates a bit of a confusion now between the distinction between an "Artillery Observer" and the existing "Artillery Officer" role. In a future update we should figure out what to do with the off-map officer role; be it combining the functionality or changing it so that the artillery officer can call in air-strikes or something similar.