Putot update: Most of the 1st visual pass done Some optimization passes (TR, texture usage, and network) Fixed various statics, textures, and added a few assets
Added a "summer" variant to the obstacle "minefield" texture for use on normandy maps
Made a crappy static in StreetFurn. use a smaller texture for a small part on the static

IMPORTANT NOTE:
I've not included changes to DH_Ground_tex.utx so Putot will not load immediately after this commit.  This was because Kashash made changes to it and I need to get the new version before I can merge.