Much needed weapon buffs and tweaks: * Reduced recoil values on PPSH41, PPS43, PPD40, MP40, Sten, and STG44 * Decreased accuracy on MP40, PPSH41, and PPD40 * Improved Sten, M1 Carbine, and PPS43 accuracy * Slightly improved STG44 and G43 accuracy * Minor tweak to BAR recoil, lengthening the rate of the recoil
The outcome of this update should be:
A buffed Sten Gun (less recoil and more accurate)
A buffed STG44 (should be stronger as an assault weapon)
Buffed Soviet SMGs (much less recoil, but generally less accurate)
All in all the Soviet SMGs will feel more like other DH SMGs and less like RO weapons

The MP40:
MP40 will seemingly be better and should be more easy to handle, but is less accurate now
The MP40 was able to put down accurate fire when rested, which perhaps broke it out of its role as assault and allowed for some effectiveness at ranges it should not be.  The decreased accuracy will force the MP40 into the role it should be.

The PPS43:
PPS43 got a big buff, but it was probably the worst of the three Soviet SMGs.

The STG44:
The STG44 was just as effective in the assault role as other nation's LMGs, perhaps worse as it has less damage than Bren and BAR.  The STG44 should be in a class of its own, it is not a LMG.  These changes should make it noticeably better in the assault role and these changes should be significant enough to make that difference apparent.  Another reason for this change is upon reviewing values, I noticed that the STG had one of the highest horizontal recoils, higher or equal to most rifles.