More work on setup phase manager:  * Added variable in GRI to state if game is in setup phase or not  * Added check in CanMantle() in DHPawn to prevent mantling while in setup phase
Note on the purpose of this:
The mantling system is not perfect and players were able to bypass the collision of the black box boundary during setup phase. I tried thickening the collision to prevent it, but once the player is considered in the collision they are free to move. I assume its how the collision on DSMs is handled, but for now a work around fix is to just straight up prevent mantling while in setup phase. Its not like mantling is needed while in the seutp phase anyways.