Varaville update fixed 2nd QA issues:
Changes:
Deleted obsolete textures to decrease file size.
Added 2 static vehicles with headlights on for ambient effect. (looks pretty good.)
Replaced dodgy hedges.
Remade the forest to get a more realistic feel to it.
Increased the density of the decolayer and made the grass less tall.
Added a couple of ambient sounds.