Small Varaville update:
Spiced up the grass by added SM grass around the map.
Changed textures on the thick hedges so they look darker.
Changed the last cap so they are recappable by axis to simulate a counter attack.
Set a cull distance of 10000 to all outside SM's (fog range is 9700)
Changed fences to reduce the use of textures.