Reverted the earlier changes made to the DHSpawnPointBase that tried to remove some redundancies in the passing of the GRI as an argument. Unfortunately this was put in for a reason that I had forgotten (not to self, be suspicious of Matt's suggestions!)
The GRI was unavailable to net clients after the change. In theory we could put an `AllActors` call to fetch the GRI on the client for each spawn point but that would just add lots of extra execution time and there'd be no guarantee that the GRI would even be present on the client at that time.

Reverted all those commits. All is well with the world again.