Significant changes to the voice system:  * Removed limit on VoiceID to be less than 32 (will break bit field masks, so muting players, local, and public channels will no longer work)  * Replaced "Team" channel with "Unassigned"  * Replaced "Public" channel with "Command"  * Hijacked "Local" channel so it is no longer used and instead when a player "Speaks Local" they speak in their own private voice channel  * Changed the IsMember() function to handle new channels appropriately  * Change some default controls to fit the changes
Note how it all works:

The "Team" channel is now just labeled as "Unassigned" and it basically acts the same way as it did before, but now it will ignore players that are in a squad.  Also I've made it so if you are in a squad, you can't "Speak Unassigned" which if allowed would be confusing (speak but not receive).

The "Public" channel is now just labeled as "Command" and it was changed to act the same way as "Team" was, but now it will ignore players that are not Squad Leaders.  Also I've made it so only Squad Leaders can "Speak Command"

Now for the crazy part!
The local channel is FUBAR, so lets just ignore it!  Instead I've made it so if you "Speak Local" instead of activating Local it activates the players personal voice channel.  Which always existed before!  However, now its more useful in that when players speak into their own channels, it does a member check and anyone on their team and within a radius can hear them!  Functionally it is the same as "Local" was before, but now it shouldn't break with more than 32 players because it isn't bit mask based.