Changes to default and control profiles:  * Changed how the system works a bit  * New default control scheme  * Added RO Classic profile (pretty much exactly the same as the previous default controls)  * Added Pro 104 profile (mean for the windows 104 (US Standard) keyboard layout, very similar to the previous "Recommended" profile, but has some changes)  * Added Pro 105 profile (meant for the Windows 105 keyboard layout often used in the UK)
Notes:
There was significant key mapping done here, so it is likely I made a mistake.  The defaultuser.ini needs update still which will basically map out the same keys as the new default controls.  We need to discuss keys to use for the deploy smoke commands and vehicle lock command.