Setup system for attrition to begin taking away reinforcements after a certain amount of time passes, defined as a curve.
This solves multiple issues:
A) It solves the problem with low player counts, the round will never end because there aren't enough players, now it will
B) It makes it difficult to have reinforcements too high at start
C) Because it takes away from both teams equally the outcome is as fair as it can get a timer might bring up issues with who really won
D) With no timer, players shouldn't be pushed/rushed, like the level is going to end at a very specific moment
E) Because A is solved, it removes the need to scale everything down for less than ideal players

So now we don't need to put in a time for the level.  I hope it works out.