Added smoke launchers to all relevant vehicles.
British 2 inch smoke bomb thrower in: Cromwell and all Sherman variants (will also be in Churchill).
I think probably plausible and realistic for all Shermans to have them by our stage of the war.
The US manufactured their own version of the weapon and they fitted this to all Shermans built or sent back to the tank depots from around mid-1943.
The British also retro-fitted the weapon to some (not sure what proportion) of their earlier Shermans, especially during Firefly conversions.
Not sure if the US also retrofitted any, as it was listed as standard equipment and the US did a lot of upgrading to standard when preparing for Overlord.
All M4A3 and the M4A1 76mm are all late war production versions, so they would have the bomb thrower as standard, and the Firefly would probably have it as part of the standard conversion.
That just leaves the M4A1 75mm as an earlier model, but would still have had the weapon if built after mid '43 or if retro-fitted, perhaps more likely on the Brit version.
Currently all the bomb throwers are range adjustable, but early versions of this weapon had a fixed range, so I want to consider whether some earlier vehicles should have that version with the range adjustment removed (easily done).
I believe the US manufactured their own 'M3' version of the weapon, which was the range adjustable type, so if that's correct the only non range adjustable versions would be early British retro-fits.
Note you don't often see the launcher hole in real photos of surviving Shermans, because the US removed and plugged them immediately after the war.

The German 'nahverteidigungswaffe' (close defence weapon) in: late StuG, StuH, jagdpanzer IV, jagdpanther, jagdtiger, panzer IV ausf J, panther ausf G, 'late' tigers (nearly all our versions are late) and king tiger.
Removed it from the early StuG and haven't added it to the panzer IV ausf H or the panther ausf D.

Set up each vehicle smoke launcher fire position as accurately as possible.
Nearly all German vehicles have the firing port and plate modeled, either in the mesh or at least the model skin.
Most of the allied vehicles don't have the bomb thrower hole (or surrounding metal comb/lip the Brits tended to add), but I have positioned it as accurately as I can from photos.
The jumbo does very helpfully have it clearly shown in the turret skin.
Added a TODO to probably delete InitialAmmo from the DHVehicleSmokeLauncher class and instead set NumSmokeLauncherRounds in the cannon's def props, as no. of rounds carried is likely to vary from vehicle to vehicle.