Updated display settings tab:  * Now has extended options for "Decal Stay" (Low, Normal, High, Higher, Very High, Highest)  * Decals can stay for longer than before with Higher, Very High, and Higest
Notes:

A high setting did and will still sets the DecalStayScale to 2.0, which would make decals stay for roughly 2-4 seconds.
Low       : 0.0 (Off, trying to rename it to that is well not worth the effort)
Normal    : 1.0
High      : 2.0
Higher    : 4.0
Very High : 8.0
Highest   : 16.0 (at this setting decals stayed around for roughly 25 seconds)

Currently in live any user can modify this "DecalStayScale" value in their ini, but likely no one knew this was possible.  Raising it is semi-dangerous, but after playing several maps without any negative impact what so ever, I think a reasonably higher value is fine.

Setting it too high is dangerous only because there seems to be an insane performance hit when viewing MANY decals in very very close proximity.  I was able to view hundreds, thousands even, without any performance drop at a distance, but up close just 60 would cause major FPS drop.  Even with a value of 2.0 shooting an MG straight in front of you will cause a major performance hit.  So its sorta important to not do that anyways.  Obviously, it can be worse with higher values like 16.0. But with some responsibility (like not shooting an MG a foot in front of you over and over) the problem just isn't there. My only remaining concern is that if for some reason you accidentally shoot a window frame and didn't know it and decals built up, the performance hit would remain for longer (instead of 2-4 seconds, it'd be more like 25 seconds). However, this is an optional thing and can easily be turned back down.

Obviously we could turn this setting into a slider or something to let the user set even higher values, but I see no point.  Because there really isn't a point in turning the value that much higher, because looking away (for a bit) or dieing or leaving the area (having them culled via anti-portal) or anything that will cull them will delete them.  So if you aren't aware of say a battle taking place in a village and drive over to the village, you won't see any decals that the battle created.  Making the decals unhelpful in the sense of "telling a story" as it is too unreliable.

However having the decals stay around for 20 more seconds or so is pretty cool and adds to the atmosphere a bit.