Major refactor of `DHScoreboard`.
The biggest part of this is that the team rendering logic has been refactored out into it's own function `DHDrawTeam`, which saves on hundreds of lines of duplicate code and will make it easier to refactor in the future.

Also removed the bit of functionality that displayed the "Objectives Held" for each team (took up a whole line, and I never even realized it was there, to be honest).