Fixes to Admin
@MattUK, extending the INI version of AccessControl made Admin stop working.  RO did not use the INI version, it does not work.  In the case you did test and know something I don't to get it working, please fix it so no additional steps are necessary and revert this commit.  Otherwise, please test in the future, as I spent 4 hours trying to figure out why I couldn't use admin.

Also your commit which removed the force change to the DHAccessControl in DarkestHourGame was needed for servers that have yet to update their INI files to use the DH class, which is why I put it there.  At a later point I was gonna remove that, perhaps before release.  But was handy in the stance that if a server doesn't update their INI, we'd still have devlogin() functionality working.