Changed the render modes in the settings page, replacing software mode with openGL.
Software does not work and setting it will prevent the game from launching causing players to have to ask how to get their game to launch again.  Added in tips to inform which is recommended, which has issues, or is unstable.  OpenGL is not stable on windows, it will crash and crash often, but is a better option than software.