Fixed a whole bunch of bugs, namely https://github.com/DarklightGames/DarkestHour/issues/771 and *possibly* https://github.com/DarklightGames/DarkestHour/issues/768.
The underlying bug here is that changing to spectator was broken in that it didn't actually "leave the team" or run any of the usual "you left the team" type logic (clearing squad rally points etc.); it essentially just set the `bIsSpectator` to `true` and only *appeared* to work because of incorrect logic in other parts of the code.

Fixing *that* revealed that the way team counts were displayed and calculated were broken as well. RO (and us by extension) were using a counting method that *excluded* players who hadn't yet selected their role. I have no idea why this was done, as it that logic was subtly breaking the team-balancing logic as well as the display of the team counts in several places. This was changed back to simply use the underlying `TeamInfo.Size` variable that UT2K4 uses which, as an added bonus, actually results in a minor optimization.

Also fixed a very minor bug where the game wouldn't prompt you to select a role if you had selected a team but not a role and were flying around in spectator mode.