Partial fix for https://github.com/DarklightGames/DarkestHour/issues/782. The animations (stilted as they are) will now consistently play in third-person.
There was a couple of problems. The first problem is that the `AllowFire` function was failing on the server even when it succeeded on the client, so the `ThirdPersonEffects` event would never get called. To fix this, I switched to a method that would work on both the client and the server. In doing this, I also effectively made the system server-authoritative instead of the client-authoritative hack-job that it was before.

Upon doing this, I noticed that there was a very strange bug occurring where the `ThirdPersonEffects` would fail to be called if the following steps were performed:

1. On one client, press down the fire button to start digging a construction.
2. While the dig animation is still happening, un-press fire buton and then hold it down.
3. Keep it held down and observe client through another client.

The observing client that is watching the digging client will *not* see subsequent digging animations so long as the button is held down or if the steps are repeated. As a hacky and quick fix to solve this, I set `bWaitForRelease` to `true`. Unfortunately, this means that players are unable to simply hold down the dig button until the job is done and, instead, must click every time. This is acceptable for a beta release, but it would be nice if we could get another set of eyes on this to fix this particular problem.