Fix for exploit https://github.com/DarklightGames/DarkestHour/issues/822.
There are two new variables in `DHConstruction`, `bCanBeTornDownByFriendlies` and `FriendlyFireDamageScale`. The Platoon HQ has these variables set to `false` and `0.0`, respectively, effectively making the Platoon HQ immortal to friendly fire and tear-down damage.

Also moved `BroadcastTeamLocalized` to a static function `DarkestHourGame`, since this function is duplicated in a number of places.