Added a weapon accuracy modifier a server can change. This is mainly for testing, but the system could be useful for other purposes.
How it works:
It works by creating a variable in the pawn, which will affect the spread in all DHProjectileFire classes. The variable in pawn though is set when it spawns to the value in DHGame. Admins can change the modifier with a set command (ex- admin set DH_Engine.DarkestHourGame AccuracyModifier 2.0)

1.0 value is no change, 2.0 doubles the spread value, 0.5 halfs.

Any pawn that spawns after the value is changed will use the new value.