Fix for https://github.com/DarklightGames/DarkestHour/issues/838 (armour-piercing bullets (i.e. 50 cal in game) cause 'hit side detection' bugs on turrets and other vehicle weapons).
The problem arose because of what I have found to be a HUGE difference between a high speed projectile's location when the native Touch() event gets called and the actual hit location. Tests showed that a projectile can be UP TO 10M beyond the hit location by the time Touch() is called!!! This can be in single player at normal game speed, not because of network lag or a speeded up game. Touch() remedies this by doing a TraceThisActor() line trace on the hit actor, just to get an accurate hit location, which it then passes into ProcessTouch(), which is the 'business end' of Touch().

The bug affecting AP bullets arose because in bullet classes I was calling PenetrateVehicleWeapon() and PenetrateVehicle() check functions, but without passing in the HitLocation, so they reverted to using the projectile's location.

Rather than passing around HitLocation, I have simply deleted those two check functions and moved the simple, one line checks into ProcessTouch() and HitWall(), which is the other event that handles hits on vehicles (on the vehicle base, not weapons). Having them as separate functions served no purpose. I think when I wrote it ages ago I had in mind that the future AP bullet class would extend the bullet class, which would have allowed simple overriding of the two short check functions. But it didn't turn out that way and the AP bullet class extends the shell class.