Significant change to the way that "game types" (ie. Push, Advance etc.) are handled. We were using an enumeration to set the type (eg. GT_Cutoff, GT_Advance), however this doesn't scale particularly well and isn't conducive to modding.
This change makes the game types their own classes (e.g. DHGameType_Advance) and extract all the gameplay-affecting functionality flags out of `DarkestHourGame` and into `DHGameType`. This may seem like a trivial change, but it opens the door for easier modding and compartmentalization of functionality in the future. For example, we could have a `RoundEnded` hook that the game type responds to, which would stop the `DarkestHourGame` class from continually growing when we want to modify functionality in a specific game type. This is also the perfect time to do this because there are no custom non-push maps out in the wild currently.

As an added bonus, the game type names are all localized rather than hard-coded (DON'T HARD CODE STRINGS!)

Existing non-push maps will need to have their `GameType` updated in `DH_LevelInfo`, but that will come in a later commit since I know that many maps are still being worked on.