Control Profiles Update:  * Setup new variables for control profiles, so the profiles could be better formatted and understood in script  * Added the new default controls (vehicle locking and swapped the vehicle smoke range controls as they were backwards)  * Added some unbinds for the JoyStick buttons to the added default profile  * Added a XBox 360 controller profile along with a new `Axis` for firing which allows the controller's right trigger to work
Notes:
We do not officially support controllers (Joysticks) for DH, but I thought I'd make it somewhat easy to setup for a player who needs to play on one for health reasons. Its really about convenience to the player and to help those that do already play with a controller to have slightly better support.

Because of the new aBaseFire Axis is not natively supported, it spams the log when in interface menus with a "bad Axis command" message. This is not harmful in anyway, but annoying to us devs.

The new xbox controller profile is meant to be played mostly on the controller, but will require some mouse/keyboard commands due to how complex DH controls are (mostly to do with interface menus and chatting). It will also conflict with the LeftMouse fire, unbind JoyZ if you wish to use LeftMouse to fire as otherwise it'll be very inconsistent. `set input JoyZ` to unbind and of course you'll need to bind fire to a non-Axis button on the controller then.

The controller seems to have issues starting/stopping vehicle engines, this may be solvable in the future, but for now its a known bug.

Finally there seems to be some issues exiting vehicles with the controller, the user many have to hit the "Use" button on the controller many times before it works.