Major refactor of construction system that touches almost everything in it:
There were a number of static functions that were being passed context-relevant information like the team index, level info etc. The problem was that, occasionally, a new function would need to be made, or a new bit of contextual information would need to be passed in. This made the whole system rather inconsistent and difficult to follow

To remedy this, I've introduced a new `Context` struct that has all of the necessary context information in it. This unifies the calling signatures across many functions and makes the system easier to understand and extend in the future.

Though I've tested this pretty thoroughly, please be on the lookout for strange construction bugs.