Fixed a bug where vehicles (especially Allied tanks) could not be resupplied at a logistics truck (DHResupplyAttachment). Also made it run the radius check run on the CollisionRadius value instead of a floating 300.0 value.
Notes:
It was doing a visible check, which world geometry (terrain) can "trace stop". So the trace to the 'bone origin' of the collision box, which is the bone the collision box is attached to fails. In most cases Allied vehicle's `collision body bone` falls under the terrain, so the terrain would block the visible check. The reason why it would sometimes work (very rarely) was that the trace probably succeeded because of trace geometry detection errors.

The downside is that vehicles can "once again" resupply on opposite sides of a wall, but it is at least reliable for all our vehicles.