Tanks no longer spawn with max munition capacity!!!
This is to simulate that not every tank always happened to have max carrying capacity at the start of its engagement. It also simulates that the tank could have been part of the fighting already or relocated from another battlefield and is not at full supply. The amount of rounds reduced is mostly geared towards HE rounds, at nearly a 50% reduction across the board. This will help reduce the amount of tank HE shelling and create more situations where tanks need to resupply. AT guns took the largest hit and after "building" an AT gun a truck will likely need to resupply it shortly.

Tanks often die before they need to resupply in DH, which pretty much means there is too much supply and very minimal demand. We want there to be a "logistical" element to DH and that means shortening the initial supply to create more a demand. Eventually we will probably make these values based on the team's "supply strength." But for now, we want to see how much this affects gameplay.

Tanks can always resupply and build up their ammo to maximum capacity after spawning.