Implemented second part of task https://github.com/DarklightGames/DarkestHour/issues/864.
There is now a "chain" icon that appears on neutral-locked objectives. The capture bar also has this icon and a short text explanation ("Connected objective(s) not secured").

Fixed a very minor bug where neutral-locked objectives would actually allow one "tick" of progress to be made by an attacking team, resulting in a flashing icon on the map and general confusion.

Also re-stated the `UpdateCompressedCapProgress` function, the RO version was a crime against programming, the new one is 4 lines, can actually be understood, and is technically faster (RO version was doing a bunch of redundant checks because the person who wrote it evidently thought that assigning a variable to the same value would trigger replication, which it doesn't).