Fix for very old bug https://github.com/DarklightGames/DarkestHour/issues/352 (players who exited a vehicle can appear to have no weapon).
Got to the bottom of this and it was a really odd one, not at all what I expected!

The problem isn't that the player's 3rd person weapon attachment actor (WA) isn't being treated as network relevant and so replicated to the viewing player, or that one of its visibility properties wasn't being replicated. It's actually the opposite - when the player in the vehicle stops being net relevant and so their vehicle and player pawn actors get destroyed on the viewing player's client, the problem is the WA often DOESN'T get destroyed!

This is really odd and shouldn't ever happen. I've found that it happens nearly all the time when the player is in the driver's seat and only occasionally when he's in a vehicle weapon position. And there is a 100% correlation between the WA being or not being destroyed (checked by adding debug logging) and the WA bug manifesting, i.e. bug only and always happens when WA is not destroyed. I think the difference is occurrence rate between the driver and weapon positions isn't relevant; probably just a symptom of different network timing when the actors drop out of relevance.

The result is the WA hangs around on that client, kind of orphaned. Not torn off from the server or anything, just being there when it shouldn't and losing its proper connection and updating. Tests show that although it can't be seen, the WA stays where it was when the player was in the vehicle and doesn't move when the player moves out of and away from the vehicle. Not does it move if the vehicle is subsequently driven away.

Because the vehicle player's pawn is destroyed when he drops out of net relevancy, the WA loses its 'Base' reference, which used to be the player pawn it was attached to. Similarly it loses its Instigator and Owner references, which were also to the pawn. We cannot have an actor reference to an actor that no longer exists. The reason the WA can't be seen hanging in the air where the driver was is because a WA is bOnlyDrawIfAttached, so when it loses its Base pawn it becomes invisible.

The real problem occurs when the driving player becomes net relevant to the viewer again. Normally his WA would also become relevant at this point and be replicated to the viewer's client, along with his player pawn and vehicle actors. But if it wasn't destroyed before, it's still there now and doesn't get re-replicated. Crucially, neither are its replicated properties updated - and specifically it doesn't receive an updated Base/Instigator/Owner reference to the newly re-spawned player pawn. So without being attached to a Base actor, it stays invisible and doesn't move, just hanging around to the end of the game.

So that's what's directly causing the problem, although clearly there's an underlying, more fundamental replication problem that's causing that.

I first suspected bFastAttachmentReplication, which is an inventory attachment property, true by default, described as "only replicates the subset of actor properties needed by basic attachments whose common properties don't vary from their defaults". The class is marked as 'nativereplication' so there's some native optimisation at work on these attachments and I thought it was stopping Base etc from being replicated because it wasn't expecting them to require updating. So I tried making bFastAttachmentReplication false, seeing if it would solve the problem at the expense of some native optimisation. Seemed to work but after trying lots of times I found it only sharply reduced the incidence of the bug, pretty much to how frequently it was originally occurring when the player was in a vehicle weapon position.

So I then looked at bOnlyDirtyReplication, which is also true by default for inventory attachments - and this is the root cause. The description of bOnlyDirtyReplication says "If true, only replicate actor if bNetDirty is true. Useful if no C++ changed attributes (such as physics). bOnlyDirtyReplication only used with bAlwaysRelevant actors." It's the last bit - weapon attachments are not bAlwaysRelevant actors, so bOnlyDirtyReplication shouldn't be used in these classes. All available web documentation always confirms this. Tested it by simply making bOnlyDirtyReplication false, no other changes, and it's a 100% reliable fix.

I suppose that when handling replication for a bOnlyDirtyReplication actor, the native code is assuming the actor will always be relevant and exist permanently on every net client, so no need to consider actor destruction and re-replication. Making it false is certainly the fix and has no apparent downside, which I would not expect.

Had a check for any other classes that have bOnlyDirtyReplication but that are not bAlwaysRelevant. The only other one is the Inventory class, which is the 1st person weapon class. I'm sure bOnlyDirtyReplication could also be changed to false in that class, but a 1st person weapon is only ever net relevant to its owning player, so it perhaps never causes any problems. Worth bearing in mind though if weird problems crop up with 1st person weapons that could be caused by failure to destroy/re-replicate the actor.