Implemented "squad leader's assistant" role (https://github.com/DarklightGames/DarkestHour/issues/920)
This allows a squad leader to assign a player as the assistant to the squad leader. Assistants are able to mark enemy positions on the map and also place constructions.

This will allow the squad leader to delegate construction tasks to another player in the squad, so they don't have to be in 2 or 3 places at once.

TODO: Still need an icon to differentiate the assistants from regular members.