Significant updates to GMC, OpelBlitz, and ZiS-5V trucks:  * Added more wheel friction so "burn outs" don't happen so often  * Adjusted handbrake so it isn't so effective  * Increased health so the vehicle can survive significant small arms fire and even AT rounds  * If the AT round/rocket fails to hit the engine it will just do overall vehicle damage (takes 2 panzershreks to destroy to body)  * If the AT or small arms hit the engine, once the engine is destroyed, the vehicle is disabled and will be destroyed in 10 seconds  * Small arms (should) be able to kill occupants along with HE & AT just as before (but it should be harder to 1 shot the vehicle and kill EVERYONE)  * Lowered the damage and explosion radius  * Spike timer will now destroy an occupied vehicle if it is not treaded (wheel vehicles only)  * Spiked vehicles now pass LastHitBy for killer instead of self
Notes:
This is basically exactly what I wanted with the https://github.com/DarklightGames/DarkestHour/pull/767 but done without any extra coding or system recodes.

This is a positive change in that hits to these vehicles that should be relatively ineffective are close to that, but of course add up to make it not completely. Effective hits (to  the engine) or (to the crew) require more skill and have much more realistic results. Hitting the engine, but not the crew disables the vehicle, but the crew have time to exit and take cover.

Players who just want to be killed after their wheeled vehicle takes a critical engine blow can remain in the vehicle (optionally).

It is possible to hit the engine and still kill crew with the same round.

Doing this on select vehicles as we want to make sure the changes work well and are accepted.